backend/requirements.txt
Django
django-filter
djangorestframework
djoser
gunicorn
psycopg2-binary
Pillow
python-dotenv
PyYAML
